+++ src/proto.rs
@@ -1,14 +1,18 @@
 #![allow(rustdoc::all)]
 #![allow(clippy::all)]
 
-cfg_if::cfg_if! {
-    if #[cfg(not(target_arch = "wasm32"))] {
-        pub mod report {
-            tonic::include_proto!("report");
-        }
-    } else {
-        pub mod report {
-            include!(concat!(env!("OUT_DIR"), "/report.rs"));
-        }
-    }
+#[allow(unknown_lints)]
+#[allow(unused_attributes)]
+#[cfg_attr(rustfmt, rustfmt::skip)]
+#[allow(box_pointers)]
+#[allow(dead_code)]
+#[allow(missing_docs)]
+#[allow(non_camel_case_types)]
+#[allow(non_snake_case)]
+#[allow(non_upper_case_globals)]
+#[allow(trivial_casts)]
+#[allow(unused_results)]
+#[allow(unused_mut)]
+pub mod reports {
+    include!(concat!(env!("OUT_DIR"), concat!("/proto/reports.rs")));
 }