+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.14](https://github.com/async-graphql/async_graphql_apollo_studio_extension/compare/v3.2.13...v3.2.14) - 2024-03-07
+
+### Fixed
+- fixing reqwest in build-dependency
+
+### Other
+- updating reqwest to use rustls
+- setting default-features to false for reqwest
+- *(deps)* bump chrono from 0.4.34 to 0.4.35
+
 ## [3.2.13](https://github.com/async-graphql/async_graphql_apollo_studio_extension/compare/v3.2.12...v3.2.13) - 2024-03-04
 
 ### Fixed